// client-id/src/app/core/models/test-info-save.ts
export class TestInfoSave{
    inspectionT: InspectionSave;
    inspectionCnfrmnIssuT: CnfrmnIssuTSave;
    userId: string;
    constructor(){
        this.inspectionT = new InspectionSave();
        this.inspectionCnfrmnIssuT = new CnfrmnIssuTSave();
        this.userId = null;
    }
}
// ...
export class CnfrmnIssuTSave{
    athrzSn: number;
    athrzCnfrmnIssuSn: number;
    issuReqstNbfps: number;
    issuOutptNbfps: number;
    issuFeeAmount: number;
    issuPrposCn: string;
    outptIndictSeCode: string;
    outptIndictSeCodeName: string;
    athrzCnfrmnIssuDetailSn: number;
    athrzIssuNo: string;
    athrzAtchmnflSn: number;
    userId: string;
    ktcChargerId: string;
    ktcChargerNm: string;
    ktcChargerDeptNm: string;
    outptAt: string;
    confmnIssuObject: string[];
    athrzDetailSn: number;
    issuProgrsSttus: string;
    athrzAtchmnflCours: string;
    athrzAtchmnflNm: string;
    athrzProcessSttusCode: string;
    rgsDe: string;
    constructor(){
        this.athrzSn = 0;
        this.athrzCnfrmnIssuSn = 0;
        this.issuReqstNbfps = 0;
        this.issuOutptNbfps = 0;
        this.issuFeeAmount = 0;
        this.issuPrposCn = null;
        this.outptIndictSeCode = null;
        this.outptIndictSeCodeName = null;
        this.athrzCnfrmnIssuDetailSn = 0;
        this.athrzIssuNo = null;
        this.athrzAtchmnflSn = 0;
        this.userId = null;
        this.ktcChargerId = null;
        this.ktcChargerNm = null;
        this.ktcChargerDeptNm = null;
        this.outptAt = null;
        this.confmnIssuObject = [];
        this.athrzDetailSn = 0;
        this.issuProgrsSttus = null;
        this.athrzAtchmnflCours = null;
        this.athrzAtchmnflNm = null;
        this.athrzProcessSttusCode = null;
        this.rgsDe = null;
    }
}