// client-id/src/app/core/models/test-share-models/entrps-address.ts
export class EntrpsAddress{
    entrprsAdresSn: number;
    entrprsSn: number;
    entrprsAdresSeCode: string;
    entrprsAdresSeCodeNm: string;
    entrprsAdresNm: string;
    adresTelno: string;
    adresFxnum: string;
    adresSeCode: string;
    detailAdres: string;
    bassAdres: string;
    zip: string;
    zip1: string;
    zip2: string;
    dmstcAt: string;
    constructor(){
        this.entrprsAdresSn = 0;
        this.entrprsSn = 0;
        this.entrprsAdresSeCode = null;
        this.entrprsAdresSeCodeNm = null;
        this.entrprsAdresNm = null;
        this.adresTelno = null;
        this.adresFxnum = null;
        this.adresSeCode = null;
        this.detailAdres = null;
        this.bassAdres = null;
        this.zip = null;
        this.zip1 = null;
        this.zip2 = null;
        this.dmstcAt = null;
    }
}